+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed goto-global-line command.
 - Implemented c.recursiveImport. 
+- implemented g.execute_shell_commands_with_options.
 
 ## 1.0.5